evo-prop-crud/requirements.txt
arrow==1.3.0
asgiref==3.8.1
autopep8==2.3.1
boto3==1.18.64
botocore==1.21.64
blessed==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
django==4.2.5
django-cors-headers==4.4.0
django-environ==0.11.2
django-import-export==4.1.1
django-picklefield==3.2
django-request==1.6.3
django-silk==5.2.0
django-storages==1.11.1
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.43
gprof2dot==2024.6.6
idna==3.7
jsonfield==3.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
tablib==3.5.0
tomli==2.0.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
wcwidth==0.2.13
uvicorn